Allow URL(s) as source in jsonschema2pojo-gradle-plugin
Closes #1010
