+++ jsonschema2pojo-gradle-plugin/src/test/groovy/org/jsonschema2pojo/gradle/JsonSchemaPluginSpec.groovy
@@ -16,22 +16,18 @@
 package org.jsonschema2pojo.gradle
 
 import static org.hamcrest.MatcherAssert.*;
-import static org.hamcrest.Matchers.*;
 
 import java.lang.reflect.Field
+import java.nio.charset.StandardCharsets
 
 import org.apache.commons.io.FileUtils
-import org.gradle.tooling.BuildLauncher
-import org.gradle.tooling.GradleConnector
-import org.gradle.tooling.ProjectConnection
-import org.jsonschema2pojo.gradle.JsonSchemaExtension
 import org.junit.Test
 
 class JsonSchemaPluginSpec {
 
   @Test
   void documentationIncludesAllProperties() {
-    String documentation = FileUtils.readFileToString(new File("README.md"));
+    String documentation = FileUtils.readFileToString(new File("README.md"), StandardCharsets.UTF_8);
 
     Set<String> ignoredProperties = new HashSet<String>() {{
         add("sourceFiles");
@@ -52,24 +48,4 @@ class JsonSchemaPluginSpec {
     assertThat(missingProperties.toString(), missingProperties.isEmpty())
   }
 
-  @Test
-  void java() {
-    build("example/java");
-  }
-
-  void build(String projectDir) {
-    GradleConnector connector = GradleConnector.newConnector()
-    connector.useGradleVersion("5.6")
-    connector.forProjectDirectory(new File(projectDir))
-    ProjectConnection connection = connector.connect()
-    try {
-      BuildLauncher launcher = connection.newBuild()
-      launcher.setStandardOutput(System.out);
-      launcher.setStandardError(System.err);
-      launcher.forTasks("build")
-      launcher.run()
-    } finally {
-      connection.close()
-    }
-  }
 }
